.agent/requirements.txt
fastapi==0.115.5
uvicorn==0.30.0
watchdog==4.0.2
gitpython==3.1.43
sqlite-utils==3.37
openai>=1.0.0,<2.0.0
